splits classes into modules adds support for multiple proxies at once   - adds lsproxy, lsp and select commands   - changes prompt to show which is the currently selected proxy splits parser into core functionality and user functionality provided example user parser in customParser.py API change for commands:   - nobody but application and sockethandler contains a proxy reference   - nobody but application contains a parser reference   - application builds a new parser object with every reload   - completer is generated and owned by parser only. exit command now properly joins threads instead of forcing os._exit() listening socket now timing out, connection logic changed to allow for joining threads sockethandler now doesn't need other, parser figures that out fixes savevars command not adding new lines refactoring using 0 <= x < limit constructs and such refactor to use format strings except for string.format() refactor enum members now CAPITAL_CASE refactor "for k in dict.keys()" to "for k in dict" refactor some loops over enumerables into [f(x) for x in enumerable] constructs removed some unused variables, mostly exceptions and unpacked data from fuction returns fix bad indent in EColorSettingKey
