+++ proxy/eSocketRole.py
@@ -1,0 +1,24 @@
+from enum import Enum, auto
+
+class ESocketRole(Enum):
+    server = auto()
+    client = auto()
+
+    def __eq__(self, other) -> bool:
+        if other is int:
+            return self.value == other
+        if other is str:
+            return self.name == other
+        if repr(type(self)) == repr(type(other)):
+            return self.value == other.value
+        return False
+
+    def __gt__(self, other) -> bool:
+        if other is int:
+            return self.value > other
+        if other is str:
+            return self.name > other
+        if repr(type(self)) == repr(type(other)):
+            return self.value > other.value
+        raise ValueError("Can not compare.")
+
